refactor: 예외 처리 관련 로직 패키지 변경
