+++ backend/ddang/src/main/java/com/ddang/ddang/common/exception/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.exception.dto;
+package com.ddang.ddang.common.exception.dto;
 
 public record ExceptionResponse(String message) {
 }
