+++ backend/ddang/src/test/java/com/ddang/ddang/report/presentation/ReportControllerTest.java
@@ -25,7 +25,7 @@ import com.ddang.ddang.authentication.configuration.AuthenticationStore;
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
 import com.ddang.ddang.chat.infrastructure.exception.ChatRoomNotFoundException;
-import com.ddang.ddang.exception.GlobalExceptionHandler;
+import com.ddang.ddang.common.exception.GlobalExceptionHandler;
 import com.ddang.ddang.qna.application.exception.InvalidAnswererException;
 import com.ddang.ddang.qna.infrastructure.exception.AnswerNotFoundException;
 import com.ddang.ddang.qna.infrastructure.exception.QuestionNotFoundException;
