+++ backend/ddang/src/test/java/com/ddang/ddang/user/presentation/UserAuctionControllerTest.java
@@ -24,7 +24,7 @@ import com.ddang.ddang.authentication.configuration.AuthenticationPrincipalArgum
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
 import com.ddang.ddang.authentication.configuration.AuthenticationStore;
-import com.ddang.ddang.exception.GlobalExceptionHandler;
+import com.ddang.ddang.common.exception.GlobalExceptionHandler;
 import com.ddang.ddang.user.presentation.fixture.UserAuctionControllerFixture;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
